porting to typescript
