+++ package.json
@@ -7,11 +7,12 @@
     "node": "10.14.1"
   },
   "bin": {
-    "3dtoolkit-signal": "./index.js"
+    "3dtoolkit-signal": "./dist/index.js"
   },
   "scripts": {
-    "start": "node index.js",
-    "test": "mocha"
+    "start": "node dist/index.js",
+    "test": "ts-mocha test/**/*.ts",
+    "build": "tsc -p ."
   },
   "keywords": [
     "webrtc",
@@ -23,13 +24,21 @@
     "express-bunyan-logger": "^1.3.3",
     "passport": "^0.4.0",
     "passport-azure-ad": "^4.0.0",
-    "webrtc-signal-http-ts": "^1.7.0",
+    "ts-mocha": "^6.0.0",
     "webrtc-signal-http-capacity": "^1.2.0",
     "webrtc-signal-http-heartbeat": "^1.0.0",
     "webrtc-signal-http-peer-identification": "^1.1.1",
-    "webrtc-signal-http-publisher": "^1.1.0"
+    "webrtc-signal-http-publisher": "^1.1.0",
+    "webrtc-signal-http-ts": "^1.8.19"
   },
   "devDependencies": {
+    "typescript": "^3.5.3",
+    "@types/mocha": "^5.2.7",
+    "@types/passport-azure-ad": "^4.0.2",
+    "@types/passport": "^1.0.0",
+    "@types/node": "^12.6.8",
+    "@types/express": "^4.17.0",
+    "@types/express-bunyan-logger": "^1.3.1",
     "mocha": "^5.2.0",
     "supertest": "^3.3.0"
   }
