+++ tslint.json
@@ -1,0 +1,30 @@
+{
+  "extends": [
+    "tslint:recommended", 
+    "tslint-eslint-rules"
+  ],
+  "rules": {
+    "max-line-length": {
+      "options": [
+        120
+      ]
+    },
+    "new-parens": true,
+    "no-arg": true,
+    "no-bitwise": true,
+    "no-conditional-assignment": true,
+    "no-consecutive-blank-lines": false,
+    "object-curly-spacing": true,
+    "whitespace": [
+      true, 
+      "check-module"
+    ]
+  },
+  "jsRules": {
+    "max-line-length": {
+      "options": [
+        120
+      ]
+    }
+  }
+}
